tests/wpt/css-tests/css-writing-modes-3_dev/xhtml1/writing-mode-parsing-svg1-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Writing Modes: parsing SVG 1.1 writing-mode values</title>
<link href="kojiishi@gmail.com" rel="author" title="Koji Ishii" />
<link href="https://drafts.csswg.org/css-writing-modes-3/#svg-writing-mode" rel="help" />
<meta content="This test asserts the parser and getComputedStyle works correctly for the SVG 1.1 writing-mode values." name="assert" />
<meta content="dom svg may" name="flags" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

</head><body><div data-expected="lr" style="writing-mode: lr"></div>
<div data-expected="rl" style="writing-mode: rl"></div>

<div data-expected="tb" style="writing-mode: tb"></div>

<script>
Array.prototype.forEach.call(document.querySelectorAll("[data-expected]"), function (element) {
  test(function () {
    var actual = getComputedStyle(element).writingMode;
    assert_equals(actual, element.dataset.expected);
  }, element.title || element.getAttribute("style"));
});
</script>
</body></html>
